avoid crash when printing long longs on 32 bit platform

+++ trunk/plugin/Sources/CIdTable.cpp
@@ -202,11 +202,11 @@ struct CIdTableHelper
 						assert(inValue < docCount);
 	
 						if (inValue >= docCount)
-							THROW(("Error creating ID table: v(%d) >= inDocCount(%d)", inValue, docCount));
+							THROW(("Error creating ID table: v(%d) >= inDocCount(%d)", (uint32)inValue, docCount));
 						
 						if (map[inValue] != nil)
-							THROW(("Error creating ID table: duplicate id for document %d (%s <=> %s)",
-								inValue, map[inValue], inKey.c_str()));
+							THROW(("Error creating ID table: duplicate id for document %ld (%s <=> %s)",
+								(uint32)inValue, map[inValue], inKey.c_str()));
 						
 						map[inValue] = data.Store(inKey.c_str());
 						++n;
